refactored login and signup page into authform component
